+++ src/cli/commands/commands.js
@@ -1,5 +1,5 @@
-const program = require('./program.js')
-const scan = require('./scan.js')
+const program = require('./program')
+const ble = require('./ble')
 const bean = require('./bean')
 const install = require('./install')
 
@@ -20,7 +20,7 @@ function cmd(fn) {
 module.exports = {
   programFirmware: cmd(program.programFirmware),
   programSketch: cmd(program.programSketch),
-  startScan: cmd(scan.startScan),
+  startScan: cmd(ble.startScan),
   blinkLed: cmd(bean.blinkLed),
   readAccel: cmd(bean.readAccel),
   readConfig: cmd(bean.readConfig),
@@ -31,5 +31,6 @@ module.exports = {
   sendSerial: cmd(bean.sendSerial),
   rename: cmd(bean.rename),
   readScratch: cmd(bean.readScratch),
-  writeScratch: cmd(bean.writeScratch)
+  writeScratch: cmd(bean.writeScratch),
+  listGATT: cmd(ble.listGATT)
 }